==> kang/1_train_EDA.ipynb <==
## Santander product recommendation

### 모든 칼럼들에 대한 value 및 EDA 전처리, EDA 방향 잡기

About 'Primary customer'  
http://www.rhythmsystems.com/blog/what-is-the-difference-between-a-primary-and-secondary-core-customer

- fecha_dato(datetime) : 날짜 데이터(제대로 이해한게 맞다면 15-01-28 부터 16-05-28까지
- ncodpers(id) : 고객 번호(고유값?)
- ind_empleado(status_emplo) : 직원 상태
> A : 재직중
> B : 퇴사
> F : fillial(자식의?)
> N : 직원이 아님
> P : pasive?
- pais_residencia(residence) : 고객 거주지
- sexo(sex) : 고객 성별
- age(age) : 고객 나이
- fecha_alta(become_fh) : 고객이 은행에 first holder가 된 날
- ind_nuevo(new_or_not) : 새 고객인가 아닌가, 6개월 이내 등록했으면 새 고객
- antiguedad(seniority) : 고객 서열(월단위)
- indrel(pri_or_not) : 
> 1은 primary customer(구매의사가 충분한 고객?) 
> 99는 primary customer이다가 월말에는 primary customer가 아니게 됨
- ult_fec_cli_1t : primary customer 로써 마지막 date
- indrel_1mes(type_of_cus) : 월초의 고객 타입 
> 1 : Primary customer
> 2 : 공동 소유자
> P : 잠재 고객
> 3 : 이전 Primary 
> 4 : 이전 공동 소유자
- tiprel_1mes(relation_of_cus) :월초의 고객 관계
A : active
I : inactive
P : former customer
R : Potential
- indresi(resi_ind) : 
> S(Yes) or N(NO): 고객 거주지와 은행 위치가 **같은** 나라인가
- indext(foreign_ind) : 
> S(Yes) or N(NO) : 외국인의 출생국가와 은행 위치가 **다른가**?
- conyuemp(spouse) : 직원의 배우자 인가 아닌가 
> 1일 경우 직원의 배우자
- canal_entrada(?) : 고객이 조인하기 위해 사용되어진 채널?
- indfall : 죽은 고객?
- tipodom : 주소 타입
- cod_prov :  고객 주소(Province code)
- nomprov : province name
- ind_actividad_clie
> 1 : active customer
> 0 : inactive customer
- renta : gross income of the household(가정의 총 수입)
- segmento :
> 1 - VIP
> 2 - individuals
> 3 - college graduated
- ind_ahor_fin_ult1	Saving Account
- ind_aval_fin_ult1	Guarantees
- ind_cco_fin_ult1	Current Accounts
- ind_cder_fin_ult1	Derivada Account
- ind_cno_fin_ult1	Payroll Account
- ind_ctju_fin_ult1	Junior Account
- ind_ctma_fin_ult1	Más particular Account
- ind_ctop_fin_ult1	particular Account
- ind_ctpp_fin_ult1	particular Plus Account
- ind_deco_fin_ult1	Short-term deposits
- ind_deme_fin_ult1	Medium-term deposits
- ind_dela_fin_ult1	Long-term deposits
- ind_ecue_fin_ult1	e-account
- ind_fond_fin_ult1	Funds
- ind_hip_fin_ult1	Mortgage
- ind_plan_fin_ult1	Pensions
- ind_pres_fin_ult1	Loans
- ind_reca_fin_ult1	Taxes
- ind_tjcr_fin_ult1	Credit Card
- ind_valo_fin_ult1	Securities
- ind_viv_fin_ult1	Home Account
- ind_nomina_ult1	Payroll
- ind_nom_pens_ult1	Pensions
- ind_recibo_ult1	Direct Debit

- 스페인어로 된 columns name들을 영어로 바꾸는 건 어떨까

import pandas as pd
import seaborn as sns
import scipy as sp
import sklearn as sk
import matplotlib as mpl
mpl.use('Agg')
import matplotlib.pylab as plt
from mpl_toolkits.mplot3d import Axes3D


import numpy as np
import scipy as sp
import sklearn as sk

mpl.use('Agg')

%matplotlib inline

sns.set()

warnings.filterwarnings('ignore')

#데이터 셋 불러오기
train = pd.read_csv("train_ver2.csv")
test = pd.read_csv("test_ver2.csv")

#데이터 column, row 파악
print("shape of train data : ", train.shape)
print("shape of test data : ", test.shape)

### 칼럼별 Null 파악

train.isnull().sum()

데이터의 대부분이 null 값인 column = ** ult_fec_cli_1t, conyuemp**

### 칼럼별 데이터 파악
- 시각화가 가능한 EDA : 앞 100만개를 자른 부분을 통해 국소적인 부분만 판단
- 시각화가 불가능한  EDA : 전체 데이터를 이용

#전체 데이터 셋 이용한 데이터의 unique value  파악
try :
    for col in range(len(train.columns)) : 
        print("Number of unique value in {0}:{1}".format(train.columns[col],len(train[train.columns[col]].unique())))
except Exception as e:
    print(e)
    pass

## fecha_dato

train["fecha_dato"].value_counts()

#### Comment
fecha_dato (date) : 
- 날짜 데이터, str 형태
- pandas의 datetime으로 활용가능
- ** 사실상 카테고리로 봐도 무방할 듯**

## ncodpers

index = train["ncodpers"].value_counts().index
value = train["ncodpers"].value_counts()

check = pd.DataFrame(value, index = index)

number = 0
for i in range(17, 0, -1) :
    print("train에서 ncodper가 {}인 사람들이 test 데이터에 {}개 들어가있음.".format(i, len(test[test["ncodpers"].isin(check[check["ncodpers"] == i].index)])))

(train["ncodpers"].value_counts()).max()

#### Comment
ncodpers(id) : 
- 고객 개개인의 id로 볼 수있는 는 unique한 값이 아닌 중복되는 값이 존재 -> EDA과정에서 살펴볼 필요가 있음.
-  최대 17개까지 중복
- int 형태

## ind_empleado

train["ind_empleado"].value_counts()

#### Comment
ind_empleado(emplo_status) : 
- 칼럼 값들의 대부분이 N을 가지고 있음
- str 형태인데 왜 int64라고 뜨지?
- 이정도 차이면 직원인지 아닌지에 따라 0과 1로 나누어도 되지 않을까

## pais_residencia

train["pais_residencia"].value_counts()

#### Comment
pais_residencia(nationality) :
- 국적, CountryCode를 나타내는 듯
- 대부분의 값이 ES, 이정도 차이면 스페인인지 아닌지에 따라 0과 1로 나누어도 되지 않을까
- str 형태인데 왜 int64라고 뜨지?


## sexo

train["sexo"].value_counts()

#### Comment
sexo(sex) :
- V : 여성, H : 남성 -> 여성의 수가 130만명가량 많은 편
- str형태인데 왜 int64라고 뜨지?


## age

train["age"].value_counts()

print("age의 null 개수 : ", train["age"].isnull().sum())
print("age의 value가 ' NA' 인 것의 개수 : ", len(train[train["age"] == ' NA']))

def to_int(x) :
    if x == " NA" :
        result = None
    else :
        result = int(x)
        
    return result

#value가 ' NA'인것을 제외한 후 int로 형 변환
train["age"] = train["age"].apply(to_int)

print("age의 null 개수 : ", train["age"].isnull().sum())
print("나이가 가장 많은 고객 : ", train["age"].max())
print("나이가 가장 어린 고객 : ", train["age"].min())
print("100세 이상인 고객 : ", len(train[train["age"] >= 100]))
print("15세 미만인 고객 : ", len(train[train["age"] < 16]))

#### Comment
age(age) :
- ' NA', null값으로는 인식되지 않지만, 값이 없는 수치들 발견  
- to_int 함수를 제작
> str 형태들을 int 형으로 형변환  
> null값이나 마찬 가지인 ' NA' 들을 null 값으로 만들어줌  

## fecha_alta

train["fecha_alta"].value_counts()

#최초거래일이 가장 빠른 고객의 날짜
pd.to_datetime(train["fecha_alta"]).min()

#최초거래일이 가장 최근인 고객의 날짜
pd.to_datetime(train["fecha_alta"]).max()

#### Comment
fecha_alta(become_cus) : 
- str형태의 날짜 데이터
- fecha_dato와 함께 두 번째 날짜 데이터

## ind_nuevo

train["ind_nuevo"].value_counts()

#### Comment
ind_nuevo(new_before_6) :
- 0이 대부분 -> 대부분의 고객이 새 고객이 아님
- 실수형 데이터

## antiguedad

train['antiguedad'].value_counts()

def func_for_antiguedad(x) :
    if x == '     NA' :
        result = None
    else :
        result = int(x)
        
    return result

train['antiguedad'] = train['antiguedad'].apply(func_for_antiguedad)

#### Comment
antiguedad(seniority) : 
- 공백이 있는 형태의 str 타입 
- (-)가 있음을 발견(우선 방치)
- ** numeric 전환 후, unique value 개수가 507 -> 259개로 감소('12'와  '     12' 같은 값들이 하나로 합쳐짐) **

## indrel

train["indrel"].value_counts()

#### Comment
indrel(primary_or_not) : 
- 1 일 경우 월초 primary, 99 일 경우 월말에는 primary가 아님
- 대부분이 월초 primary 상태


## ult_fec_cli_1t

train["ult_fec_cli_1t"].value_counts()

#### Comment
ult_fec_cli_1t(last_primary) :
- 마지막으로 primary customer 였던 날짜
- 대부분이 null

## indrel_1mes

train["indrel_1mes"].value_counts()

#같은 값을 가져야 하지만, 타입이 달라 데이터가 분산되어 나옴 -> 통일
# 'P'를 제외한 모든 값들을 int로 형 변환
def func_for_indrel(x) :
    number = "0123456789"
    if type(x) == str :
        if x[0] in number :
            result = int(float(x))
        else :
            result = x
    else :
        result = x
    
    return result

train["indrel_1mes"] = train["indrel_1mes"].apply(func_for_indrel)

#### Comment
indrel_1mes(type_of_cus) : 
- 같은 값이나, 서로 다른 타입으로 되어있던 값들을 하나의 값으로 통일 시켜줌
- 1이 압도적으로 많음


## tiprel_1mes

train["tiprel_1mes"].value_counts()

#### Comment 
tiprel_1mes(relation_of_cus) : 
- 카테고리 설명에 없는 4개의 N 값 발견

## indresi

train["indresi"].value_counts()

train["indresi"].isnull().sum()

#### Comment
indresi(same_country) : 
- 특이사항 없음

## indext

train["indext"].value_counts()

train["indext"].isnull().sum()

#### Comment
- null 값이 27734로 동일한 칼럼이 많은 것을 발견

#row별 null  개수를 확인...하려했으나 어찌된 일인지 이것도 수행하지 못해서 우선 패스
null_row = train.apply(lambda x: len(train.columns) - x.count(), axis=1)

## conyuemp

train["conyuemp"].value_counts()

#### Comment
conyuemp(spouse) : 
-  배우자 여부이나, 대부분이 null

## canal_entrada

train["canal_entrada"].value_counts()

train["canal_entrada"].isnull().sum()

#### Comment
canal_entrada(channel) :
- 고객이 은행에 접촉하게 된 계기, 코드에 대한 이해는 아직 찾지 못함

## indfall

train["indfall"].value_counts()

train["indfall"].isnull().sum()

#### Comment
indfall(dead) : 
- 고객이 죽었나 살았나.
- 살아있는 사람이 대부분

## tipodom

train["tipodom"].value_counts()

train["tipodom"].isnull().sum()

#### Comment
tipodom(address) : 
- primary의 주소 타입, 전부 1
-  null값을 제외하고 전부 1, 어떻게 활용할까

## cod_prov

train["cod_prov"].value_counts()

#### Comment
cod_prov(prov_code) :
- 스페인 각 지방의 고유 번호.
- 은행과의 거리를 점쳐볼 수 있을 듯(santander bank는 스페인 전국에 3개가 있는 곳으로 추정)

## nomprov

train["nomprov"].value_counts()

#### Comment :
nomprov(name_prov)
- 스페인 각 지방 이름, 어떤 지방에 사는 고객인지
- prov_code와 중복

## ind_actividad_cliente

train["ind_actividad_cliente"].value_counts()

train["ind_actividad_cliente"].isnull().sum()

## renta

train["renta"].value_counts()

#### Comment:
renta(income) : 
- 수입, 실수형인데 중복되는 데이터가 5000개씩 되는가?
- 아라님이 하신 EDA 부분을 봐야할 듯 

## segmento

train["segmento"].value_counts()

#숫자와 텍스트가 함께 있는 형태를 숫자로만 변환
def func_for_segmento(x) :    
    if type(x) != float :
        if "1" in x :
            result = 1
        elif "2" in x :
            result = 2
        elif "3" in x :
            result = 3
    else : 
        result = None
    return result

train["segmento"] = train["segmento"].apply(func_for_segmento)

train["segmento"].value_counts()

#### Comment 
segmento(segment) : 
- 고객을 3가지로 구분

## ind_ahor_fin_ult1

train["ind_ahor_fin_ult1"].value_counts()

train["ind_aval_fin_ult1"].value_counts()

for i in train.loc[ : ,"ind_ahor_fin_ult1" : ].columns :
    print("Value count of {} : ".format(i), "\n", train[i].value_counts(), "\n", "-" * 80)

#### Comment
종속변수 값들 중에는 전처리해야할 값은 없음

#전처리한 데이터 셋  csv화
train.to_csv("train_preprocessing.csv", index = False)
