==> kang/4_QDA&LDA.ipynb <==
import pandas as pd
from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
from sklearn.metrics import *

from statsmodels.stats.outliers_influence import variance_inflation_factor

X_train = pd.read_csv("X_train.csv")
y_train = pd.read_csv("y_train.csv")

y_features = [
 'ind_ahor_fin_ult1','ind_aval_fin_ult1','ind_cco_fin_ult1','ind_cder_fin_ult1','ind_cno_fin_ult1','ind_ctju_fin_ult1',
 'ind_ctma_fin_ult1','ind_ctop_fin_ult1','ind_ctpp_fin_ult1','ind_deco_fin_ult1','ind_deme_fin_ult1',
 'ind_dela_fin_ult1','ind_ecue_fin_ult1','ind_fond_fin_ult1','ind_hip_fin_ult1','ind_plan_fin_ult1',
 'ind_pres_fin_ult1','ind_reca_fin_ult1','ind_tjcr_fin_ult1','ind_valo_fin_ult1','ind_viv_fin_ult1','ind_nomina_ult1',
 'ind_nom_pens_ult1','ind_recibo_ult1']

qda = QuadraticDiscriminantAnalysis(store_covariance = True)

X_train.head()

X_train.drop(labels = y_features + ["ncodpers"], axis = 1, inplace = True)
y_train.drop(labels = "ncodpers", axis = 1, inplace = True)

threshold = 0.5
(y_pred>threshold).astype(int)
for col in y_train.columns[1:2+1] : 
    if len(y_train[col].unique()) != 1 :
        print(col)
        model_qda = qda.fit(X_train, y_train[col])
        y_pred = model_qda.predict_proba(X_train)[:,1]
        y_pred = (y_pred>threshold).astype(int)
        print(roc_auc_score(y_train[col], y_pred))
        print("\t")
        print(confusion_matrix(y_train[col], y_pred))
        print("\t")
        print(accuracy_score(y_train[col], y_pred))
        print("\t")

model_qda = qda.fit(X_train, y_train["ind_cco_fin_ult1"])
y_pred = model_qda.predict(X_train)

lda = LinearDiscriminantAnalysis(n)

vif = pd.DataFrame()
vif["VIF Factor"] = [variance_inflation_factor(X_train.values, i) for i in range(X_train.shape[1])]
vif["features"] = X_train.columns
vif
