==> kang/5_naive_bayes.ipynb <==
import pandas as pd
from sklearn.naive_bayes import GaussianNB
from sklearn.naive_bayes import MultinomialNB
from sklearn.naive_bayes import BernoulliNB

from sklearn.metrics import *

X_train = pd.read_csv("X_train.csv")
y_train = pd.read_csv("y_train.csv")

GNB = GaussianNB()

# MNB = MultinomialNB()
# BNB = BernoulliNB()

model_GNB = GNB.fit(X_train, y_train[y_train.columns[0]])

y_pred = model_GNB.predict_proba(X_train)
